v.1.2.1.

+++ HybPhyloMaker0a_preparedata.sh
@@ -21,7 +21,7 @@
 # ********************************************************************************
 # *    HybPhyloMaker - Pipeline for Hyb-Seq data processing and tree building    *
 # *                     Script 00a - Download & prepare data                     *
-# *                                   v.1.2.0                                    *
+# *                                   v.1.2.1                                    *
 # * Tomas Fer, Dept. of Botany, Charles University, Prague, Czech Republic, 2016 *
 # * tomas.fer@natur.cuni.cz                                                      *
 # ********************************************************************************
@@ -128,12 +128,13 @@ else
 	fi
 fi
 
-if [ "$(ls -A ../workdir00_dataprep)" ]; then
-	echo -e "Directory 'workdir00_dataprep' already exists and is not empty. Delete it or rename before running this script again. Exiting...\n"
-	rm -d ../workdir00_dataprep/ 2>/dev/null
-	exit 3
+if [[ ! $location == "1" ]]; then
+	if [ "$(ls -A ../workdir00_dataprep)" ]; then
+		echo -e "Directory 'workdir00_dataprep' already exists and is not empty. Delete it or rename before running this script again. Exiting...\n"
+		rm -d ../workdir00_dataprep/ 2>/dev/null
+		exit 3
+	fi
 fi
-
 #Make folder for your data
 mkdir -p $path/
 mkdir 10rawreads
