+++ HybPhyloMaker3_processpslx.sh
@@ -22,7 +22,7 @@
 # ********************************************************************************
 # *    HybPhyloMaker - Pipeline for Hyb-Seq data processing and tree building    *
 # *                        Script 03 - Process pslx files                        *
-# *                                   v.1.2.0                                    *
+# *                                   v.1.2.1                                    *
 # * Tomas Fer, Dept. of Botany, Charles University, Prague, Czech Republic, 2016 *
 # * tomas.fer@natur.cuni.cz                                                      *
 # * based on Weitemier et al. (2014), Applications in Plant Science 2(9): 1400042*
@@ -146,10 +146,12 @@ else
 		rm -d ../workdir03/ 2>/dev/null
 		exit 3
 	else
-		if [ "$(ls -A ../workdir03)" ]; then
-			echo -e "Directory 'workdir03' already exists. Delete it or rename before running this script again. Exiting...\n"
-			rm -d ../workdir03/ 2>/dev/null
-			exit 3
+		if [[ ! $location == "1" ]]; then
+			if [ "$(ls -A ../workdir03)" ]; then
+				echo -e "Directory 'workdir03' already exists. Delete it or rename before running this script again. Exiting...\n"
+				rm -d ../workdir03/ 2>/dev/null
+				exit 3
+			fi
 		fi
 	fi
 fi
