+++ HybPhyloMaker7a_astral.sh
@@ -21,7 +21,7 @@
 # ********************************************************************************
 # *    HybPhyloMaker - Pipeline for Hyb-Seq data processing and tree building    *
 # *                       Script 07a - Astral species tree                       *
-# *                                   v.1.2.0                                    *
+# *                                   v.1.2.1                                    *
 # * Tomas Fer, Dept. of Botany, Charles University, Prague, Czech Republic, 2016 *
 # * tomas.fer@natur.cuni.cz                                                      *
 # ********************************************************************************
@@ -50,6 +50,8 @@ if [[ $PBS_O_HOST == *".cz" ]]; then
 	module add jdk-1.6.0
 	module add python-2.7.6-gcc
 	module add python27-modules-gcc
+	module add python-3.4.1-intel
+	module add newick-utils-1.6
 elif [[ $HOSTNAME == compute-*-*.local ]]; then
 	echo -e "\nHybPhyloMaker7a is running on Hydra..."
 	#settings for Hydra
@@ -62,6 +64,8 @@ elif [[ $HOSTNAME == compute-*-*.local ]]; then
 	cd workdir07a
 	#Add necessary modules
 	module load java/1.7
+	module load bioinformatics/anaconda3/2.3.0
+	module load bioinformatics/newickutilities/0.0
 else
 	echo -e "\nHybPhyloMaker7a is running locally..."
 	#settings for local run
@@ -258,10 +262,12 @@ else
 		exit 3
 	fi
 fi
-if [ "$(ls -A ../workdir07a)" ]; then
-	echo -e "Directory 'workdir07a' already exists and is not empty. Delete it or rename before running this script again. Exiting...\n"
-	rm -d ../workdir07a 2>/dev/null
-	exit 3
+if [[ ! $location == "1" ]]; then
+	if [ "$(ls -A ../workdir07a)" ]; then
+		echo -e "Directory 'workdir07a' already exists and is not empty. Delete it or rename before running this script again. Exiting...\n"
+		rm -d ../workdir07a 2>/dev/null
+		exit 3
+	fi
 fi
 
 #Add necessary programs and files
