+++ HybPhyloMaker7c_mrl.sh
@@ -21,7 +21,7 @@
 # ********************************************************************************
 # *    HybPhyloMaker - Pipeline for Hyb-Seq data processing and tree building    *
 # *                        Script 07c - MRL species tree                         *
-# *                                   v.1.2.0                                    *
+# *                                   v.1.2.1                                    *
 # * Tomas Fer, Dept. of Botany, Charles University, Prague, Czech Republic, 2016 *
 # * tomas.fer@natur.cuni.cz                                                      *
 # ********************************************************************************
@@ -157,12 +157,13 @@ else
 		exit 3
 	fi
 fi
-if [ "$(ls -A ../workdir07c)" ]; then
-	echo -e "Directory 'workdir07c' already exists and is not empty. Delete it or rename before running this script again. Exiting...\n"
-	rm -d ../workdir07a 2>/dev/null
-	exit 3
+if [[ ! $location == "1" ]]; then
+	if [ "$(ls -A ../workdir07c)" ]; then
+		echo -e "Directory 'workdir07c' already exists and is not empty. Delete it or rename before running this script again. Exiting...\n"
+		rm -d ../workdir07a 2>/dev/null
+		exit 3
+	fi
 fi
-
 
 #Add necessary programs and files
 cp $source/mrp.jar .
