+++ HybPhyloMaker7e_concatenatedFastTree.sh
@@ -21,7 +21,7 @@
 # ********************************************************************************
 # *    HybPhyloMaker - Pipeline for Hyb-Seq data processing and tree building    *
 # *                    Script 07e - concatenated species tree                    *
-# *                                   v.1.2.0                                    *
+# *                                   v.1.2.1                                    *
 # * Tomas Fer, Dept. of Botany, Charles University, Prague, Czech Republic, 2016 *
 # * tomas.fer@natur.cuni.cz                                                      *
 # ********************************************************************************
@@ -48,6 +48,7 @@ if [[ $PBS_O_HOST == *".cz" ]]; then
 	#Add necessary modules
 	module add fasttree-2.1.8
 	module add python-3.4.1-intel
+	module add newick-utils-1.6
 elif [[ $HOSTNAME == compute-*-*.local ]]; then
 	echo -e "\nHybPhyloMaker7e is running on Hydra..."
 	#settings for Hydra
@@ -61,6 +62,7 @@ elif [[ $HOSTNAME == compute-*-*.local ]]; then
 	#Add necessary modules
 	module load bioinformatics/fasttree/2.1.8
 	module load bioinformatics/anaconda3/2.3.0
+	module load bioinformatics/newickutilities/0.0
 else
 	echo -e "\nHybPhyloMaker7e is running locally..."
 	#settings for local run
@@ -145,10 +147,12 @@ else
 		exit 3
 	fi
 fi
-if [ "$(ls -A ../workdir07e)" ]; then
-	echo -e "Directory 'workdir07e' already exists and is not empty. Delete it or rename before running this script again. Exiting...\n"
-	rm -d ../workdir07e 2>/dev/null
-	exit 3
+if [[ ! $location == "1" ]]; then
+	if [ "$(ls -A ../workdir07e)" ]; then
+		echo -e "Directory 'workdir07e' already exists and is not empty. Delete it or rename before running this script again. Exiting...\n"
+		rm -d ../workdir07e 2>/dev/null
+		exit 3
+	fi
 fi
 
 #Add necessary scripts and files
