+++ HybPhyloMaker7f_concatenatedExaML.sh
@@ -10,7 +10,7 @@
 # ********************************************************************************
 # *    HybPhyloMaker - Pipeline for Hyb-Seq data processing and tree building    *
 # *                     Script 07f - ExaML concatenated tree                     *
-# *                                   v.1.2.0                                    *
+# *                                   v.1.2.1                                    *
 # * Tomas Fer, Dept. of Botany, Charles University, Prague, Czech Republic, 2016 *
 # * tomas.fer@natur.cuni.cz                                                      *
 # ********************************************************************************
@@ -130,12 +130,13 @@ else
 		exit 3
 	fi
 fi
-if [ "$(ls -A ../workdir07f)" ]; then
-	echo -e "Directory 'workdir07f' already exists and is not empty. Delete it or rename before running this script again. Exiting...\n"
-	rm -d ../workdir07f 2>/dev/null
-	exit 3
+if [[ ! $location == "1" ]]; then
+	if [ "$(ls -A ../workdir07f)" ]; then
+		echo -e "Directory 'workdir07f' already exists and is not empty. Delete it or rename before running this script again. Exiting...\n"
+		rm -d ../workdir07f 2>/dev/null
+		exit 3
+	fi
 fi
-
 
 #Check if there is already partition file (PartitionFinder was already run)
 if [ -f $path/${treepath}${MISSINGPERCENT}_${SPECIESPRESENCE}/concatenatedExaML/RAxMLpartitions.txt ]; then
